+++ cmd/ironport-demo/main.go
@@ -58,8 +58,8 @@ func main() {
 
 	go func() {
 		for ev := range srv.CompletedUploads() {
-			log.Printf("completed upload: user=%q ip=%q path=%q full=%q",
-				ev.Username, ev.ClientIP, ev.FilePath, ev.FullFilePath)
+			log.Printf("completed upload: protocol=%q user=%q ip=%q path=%q full=%q",
+				ev.Protocol, ev.Username, ev.ClientIP, ev.FilePath, ev.FullFilePath)
 		}
 	}()